pyproject.toml
[project]
name = "self-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=25.12.0",
    "chromadb>=1.3.7",
    "gradio>=6.1.0",
    "isort>=7.0.0",
    "langchain>=1.1.3",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.0.0",
    "langchain-openai>=1.1.3",
    "langchain-tavily>=0.2.14",
    "langchainhub>=0.1.21",
    "langgraph>=1.0.5",
    "langgraph-cli>=0.4.9",
    "langsmith>=0.4.59",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.15",
]
